requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.1
psycopg2-binary>=2.9
gunicorn
python-dotenv
dj-database-url
